+++ application/apps/indexer/session/src/state/source_ids.rs
@@ -1,15 +1,15 @@
 use serde::{Deserialize, Serialize};
 use std::{collections::HashMap, ops::RangeInclusive};
 pub struct MappedRanges<'a> {
-    ranges: Vec<&'a (RangeInclusive<u64>, u8)>,
+    ranges: Vec<&'a (RangeInclusive<u64>, u16)>,
 }
 
 impl<'a> MappedRanges<'a> {
-    pub fn new(ranges: Vec<&'a (RangeInclusive<u64>, u8)>) -> Self {
+    pub fn new(ranges: Vec<&'a (RangeInclusive<u64>, u16)>) -> Self {
         Self { ranges }
     }
 
-    pub fn source(&self, line: u64) -> Option<u8> {
+    pub fn source(&self, line: u64) -> Option<u16> {
         self.ranges.iter().find_map(|(range, source_id)| {
             if range.contains(&line) {
                 Some(*source_id)
@@ -22,15 +22,15 @@ impl<'a> MappedRanges<'a> {
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct SourceDefinition {
-    pub id: u8,
+    pub id: u16,
     pub alias: String,
 }
 
 #[derive(Debug)]
 pub struct SourceIDs {
-    pub sources: HashMap<u8, String>,
-    pub map: Vec<(RangeInclusive<u64>, u8)>,
-    pub recent: Option<u8>,
+    pub sources: HashMap<u16, String>,
+    pub map: Vec<(RangeInclusive<u64>, u16)>,
+    pub recent: Option<u16>,
 }
 
 impl SourceIDs {
@@ -42,19 +42,19 @@ impl SourceIDs {
         }
     }
 
-    pub fn add_source(&mut self, alias: String) -> u8 {
-        let key = self.sources.len() as u8;
+    pub fn add_source(&mut self, alias: String) -> u16 {
+        let key = self.sources.len() as u16;
         self.sources.insert(key, alias);
         key
     }
 
-    pub fn get_source(&mut self, alias: String) -> Option<u8> {
+    pub fn get_source(&mut self, alias: String) -> Option<u16> {
         self.sources
             .iter()
             .find_map(|(key, val)| if val == &alias { Some(*key) } else { None })
     }
 
-    pub fn is_source_same(&self, source_id: u8) -> bool {
+    pub fn is_source_same(&self, source_id: u16) -> bool {
         if let Some(id) = self.recent {
             id == source_id
         } else {
@@ -62,7 +62,7 @@ impl SourceIDs {
         }
     }
 
-    pub fn source_update(&mut self, source_id: u8) {
+    pub fn source_update(&mut self, source_id: u16) {
         let changed = if let Some(id) = self.recent {
             id != source_id
         } else {
@@ -73,11 +73,11 @@ impl SourceIDs {
         }
     }
 
-    pub fn get_recent_source_id(&self) -> u8 {
+    pub fn get_recent_source_id(&self) -> u16 {
         if let Some(id) = self.recent {
             id
         } else {
-            self.sources.len() as u8
+            self.sources.len() as u16
         }
     }
 
@@ -91,7 +91,7 @@ impl SourceIDs {
             .collect::<Vec<SourceDefinition>>()
     }
 
-    pub fn add_range(&mut self, range: RangeInclusive<u64>, source_id: u8) {
+    pub fn add_range(&mut self, range: RangeInclusive<u64>, source_id: u16) {
         self.map.push((range, source_id));
     }
 
@@ -105,7 +105,7 @@ impl SourceIDs {
                         || requested.contains(range.start())
                         || requested.contains(range.end())
                 })
-                .collect::<Vec<&(RangeInclusive<u64>, u8)>>(),
+                .collect::<Vec<&(RangeInclusive<u64>, u16)>>(),
         )
     }
 }
